Switch from floats to doubles for the grids
* Avoids jittering problems when close to SSB

+++ modules/base/rendering/grids/renderablegrid.h
@@ -55,7 +55,7 @@ public:
 
 protected:
     struct Vertex {
-        float location[3];
+        double location[3];
     };
 
     ghoul::opengl::ProgramObject* _gridProgram = nullptr;
